gateway/requirements.txt
fastapi>=0.115,<1.0
pydantic-settings>=2.4,<3.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
python-jose[cryptography]>=3.3,<4.0
structlog>=24.4,<25.0
slowapi>=0.1.9,<0.2
